Add testing for Python 3.13

+++ eqcorrscan/utils/mag_calc.py
@@ -146,12 +146,12 @@ def calc_b_value(magnitudes, completeness, max_mag=None, plotvar=True):
     >>> b_values = calc_b_value(magnitudes, completeness=np.arange(3, 7, 0.2),
     ...                         plotvar=False)
     >>> round(b_values[4][1], 1)
-    1.0
+    np.float64(1.0)
     >>> # We can set a maximum magnitude:
     >>> b_values = calc_b_value(magnitudes, completeness=np.arange(3, 7, 0.2),
     ...                         plotvar=False, max_mag=5)
     >>> round(b_values[4][1], 1)
-    1.0
+    np.float64(1.0)
     """
     b_values = []
     # Calculate the cdf for all magnitudes
